requirements.txt
asgiref==3.8.1 ; python_version >= "3.12" and python_version < "4.0"
django-countries==7.6.1 ; python_version >= "3.12" and python_version < "4.0"
django-dotenv==1.4.2 ; python_version >= "3.12" and python_version < "4.0"
django==5.1 ; python_version >= "3.12" and python_version < "4.0"
pillow==10.4.0 ; python_version >= "3.12" and python_version < "4.0"
sqlparse==0.5.1 ; python_version >= "3.12" and python_version < "4.0"
typing-extensions==4.12.2 ; python_version >= "3.12" and python_version < "4.0"
tzdata==2024.1 ; python_version >= "3.12" and python_version < "4.0"
